LPS-171127 Extract search-spring.xml
