--- portal-impl/src/META-INF/search-spring.xml
@@ -1,17 +1,0 @@
-<?xml version="1.0"?>
-
-<beans
-	default-destroy-method="destroy"
-	default-init-method="afterPropertiesSet"
-	xmlns="http://www.springframework.org/schema/beans"
-	xmlns:util="http://www.springframework.org/schema/util"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-3.0.xsd http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-3.0.xsd"
->
-
-	<!-- Util -->
-
-	<bean class="com.liferay.portal.kernel.search.facet.MultiValueFacetFactory" id="com.liferay.portal.kernel.search.facet.MultiValueFacetFactory" />
-	<bean class="com.liferay.portal.kernel.search.facet.RangeFacetFactory" id="com.liferay.portal.kernel.search.facet.RangeFacetFactory" />
-	<bean class="com.liferay.portal.kernel.search.facet.SimpleFacetFactory" id="com.liferay.portal.kernel.search.facet.SimpleFacetFactory" />
-</beans>
