+++ portal-impl/src/portal.properties
@@ -783,7 +783,6 @@
         /META-INF/poller-spring.xml,\
         /META-INF/prefs-spring.xml,\
         /META-INF/repository-spring.xml,\
-        /META-INF/search-spring.xml,\
         \
         /META-INF/counter-spring.xml,\
         /META-INF/mail-spring.xml,\
